Finish basic database schema
We may need to add additional info to Thesis as we figure out how to
integrate with numpy (MLDP-25) - and that will add significantly to the
scope/hours of MLDP-25 while subtracting from the estimate here.
But this covers the metadata we know we need, and sets us up to start
work on MLDP-24.