Add inferred vectors to thesis instances.
When comparing theses that were in the neural net, we can use the
label. For those that were not, we need to use inferred vectors.

This strategy actually won't work longterm - we'll need to retrain the
net on everything, using our selected parameters - because we need all
the documents to be in the universe in order to power the similarity
functions. So we'll want to back out of this at some point. But it
will do for now, since it allows further datavis development.