Don't deploy neural net trainer, just models
The neural net training machinery is only meant to run locally, not on
Heroku. In fact, it fails to deploy on Heroku because we have
.gitignored the files directory in the interest of reducing size, which
causes the import of ModelTrainer to fail.

We DO want to store this machinery in version control. Therefore, we
add hamlet/neural to the .slugignore.