Try using a model that has deleted temporary training data
This uses delete_temporary_training_data with keep_inference=False
(but doctags true, because we need to use them to match up with theses,
unless we write some kind of data migration to store doctags on the
Thesis model). This should use much less memory, so maybe Heroku won't
crash.

It isn't adequate to power any of the "upload your thesis" functions,
but we can use a model trained on a much smaller data set for that.