See if we can get Travis to work
For the app to build, it needs to be able to find a neural net to use.
Our production net is kept out of version control because it's too
big...but that means Travis can't see it. This commit adds a smaller
test model that Travis can use, and moves MODEL_FILE into child settings
files so as to avoid trying to load a file that isn't present.