Iterate through lines of uploaded file
Iterating through byte chunks may split in the middle of words, or
worse, the middle of multi-byte characters. Django's file object allows
iterating by line which should be safer. Additionally, this explicitly
decodes from UTF-8. This is probably fine for now, but we'll want to add
encoding detection at a later point.