Improve citation classifier accuracy
It turns out the accuracy of the underlying extractor was actually
quite good, and we didn't need the targets system. Extra-long
candidate citations were parser garbage, but the shorter ones looked
good, and there's a wide separation between them (<300 characters for
good citations, >1500 for bad ones), so we can be quite loose on our
good/bad classifier.