Add Sentry exception monitoring
This installs Sentry for handling uncaught exceptions. In order to
actually enable it, you will need to create a project in Sentry and add
the DSN it provides to the SENTRY_DSN environment variable. This should
only be done in the production environment. If no SENTRY_DSN env var is
defined, Sentry won't report exceptions.