Converging on restored app??
The deploy is still failing, but it is actually getting to the point
where it unzips the app. It may just be failing on not being able to
move the .bkp file that doesn't exist. Let's take that out, ensure
some other commands succeed, and try again.

On EB, have also configured the environment to deploy when health
is degraded, in hopes of forcing it to accept the latest configuration;
the last successful deploy can no longer succeed, so we don't want to be
stuck on it.