What if we load wsgi
We no longer need the bit that destroys obsolete ssl conf -- logs
indicate we're getting past that part of the deploy. However, apache is
hanging. But, uh, maybe that's because we deleted the part where we
configured wsgi. Let's put it back, but move it all to port 80, since
port 443 isn't going to work until we have certbot.