Upgrade gensim
I can't get the fast C binding stuff working on Amazon Linux, so the
features involving processing uploaded text fail (they rely on a code
path that only exists with the fast C backing). I'd still like to get
it working for the sake of speed, but in the meantime, I'd like the
feature to work at all.

I tried the next minor version of gensim (3.3.0) but it failed due to
the bug fixed in this PR, which was incorporated into 3.5.0:
https://github.com/RaRe-Technologies/gensim/pull/1974

So I'm using the lowest gensim version which passes the tests. It's way
out of date now but I don't want to open up the upgrade can of worms.